requirements.txt
rich
pytest
requests
pillow
argparse
rich-argparse